+++ docs/web/docs/explanations/architect_api.md
@@ -109,4 +109,4 @@ The `useMephistoTask` hook is responsible for allowing a worker to connect to a 
 
 The `useMephistoLiveTask` hook is responsible for the rest of the packets. Data packets should be sent via `sendData` and handled with the `onLiveUpdate` callback. So long as your data is json-serializable, you can send anything you want this way.
 
-We also provide a `useMephistoRemoteQueryTask` hook, which is a wrapper around `useMephistoLiveTask` that instead allows for making remote procedure calls from static tasks (when combined with the `RemoteQueryBlueprint` or a similar API). Here people can make requests to the backend from an otherwise static task, and potentially receive responses and take action on them if they've registered callbacks. The only interface here is thus `makeRemoteCall`.
+We also provide a `useMephistoRemoteProcedureTask` hook, which is a wrapper around `useMephistoLiveTask` that instead allows for making remote procedure calls from static tasks (when combined with the `RemoteProcedureBlueprint` or a similar API). Here people can make requests to the backend from an otherwise static task, and potentially receive responses and take action on them if they've registered callbacks. The only interface here is thus `makeRemoteCall`.
