+++ examples/remote_procedure/mnist/README.md
