+++ examples/remote_procedure/mnist/__init__.py
