+++ examples/remote_procedure/mnist/hydra_configs/conf/launch_with_local.yaml
@@ -1,6 +1,6 @@
 #@package _global_
 defaults:
-  - /mephisto/blueprint: remote_query
+  - /mephisto/blueprint: remote_procedure
   - /mephisto/architect: local
   - /mephisto/provider: mock
 mephisto:
@@ -13,10 +13,10 @@ mephisto:
   task:
     allowed_concurrent: 1
     task_name: remote-query-test-task
-    task_title: "Test a static task that also has access to live backend queries"
+    task_title: "Provide feedback on our MNIST model"
     # NOTE you'll want to update your task description
-    task_description: "You will be shown a form that you can submit at any time, but can also toggle a backend query"
+    task_description: "You will draw digits. Try to fool our MNIST model, and provide us the correct label."
     # NOTE set a reasonable reward
     task_reward: 0.05
     # NOTE will want real tags
-    task_tags: "test,task,fix-me"
+    task_tags: "mnist,drawing,models,correction"
