+++ examples/remote_procedure/mnist/model.py
