+++ examples/remote_procedure/mnist/webapp/.babelrc
