+++ examples/remote_procedure/mnist/webapp/package.json
