+++ examples/remote_procedure/mnist/webapp/src/app.jsx
@@ -8,13 +8,17 @@
 
 import React from "react";
 import ReactDOM from "react-dom";
-import { BaseFrontend, LoadingScreen } from "./components/core_components.jsx";
+import {
+  BaseFrontend,
+  LoadingScreen,
+  Instructions,
+} from "./components/core_components.jsx";
 
 import {
   MephistoContext,
-  useMephistoRemoteQueryTask,
+  useMephistoRemoteProcedureTask,
   ErrorBoundary,
-} from "mephisto-remote-query";
+} from "mephisto-task";
 
 /* ================= Application Components ================= */
 
@@ -27,8 +31,8 @@ function uuidv4() {
   });
 }
 
-function RemoteQueryApp() {
-  let mephistoProps = useMephistoRemoteQueryTask({});
+function RemoteProcedureApp() {
+  let mephistoProps = useMephistoRemoteProcedureTask({});
 
   let {
     blockedReason,
@@ -36,15 +40,14 @@ function RemoteQueryApp() {
     taskConfig,
     isPreview,
     previewHtml,
-    initialTaskData,
     isLoading,
     handleSubmit,
-    remoteFunction,
+    remoteProcedure,
     isOnboarding,
     handleFatalError,
   } = mephistoProps;
 
-  const handleRemoteCall = remoteFunction("handle_with_model");
+  const classifyDigit = remoteProcedure("classify_digit");
 
   if (isOnboarding) {
     // TODO You can use this as an opportunity to display anything you want for
@@ -60,13 +63,7 @@ function RemoteQueryApp() {
     return <LoadingScreen />;
   }
   if (isPreview) {
-    if (!taskConfig.has_preview) {
-      return <TaskPreviewView description={taskConfig.task_description} />;
-    }
-    if (previewHtml === null) {
-      return <div>Loading...</div>;
-    }
-    return <div dangerouslySetInnerHTML={{ __html: previewHtml }} />;
+    return <Instructions />;
   }
 
   return (
@@ -74,8 +71,7 @@ function RemoteQueryApp() {
       <MephistoContext.Provider value={mephistoProps}>
         <div className="container-fluid" id="ui-container">
           <BaseFrontend
-            taskData={initialTaskData}
-            handleRemoteCall={handleRemoteCall}
+            classifyDigit={classifyDigit}
             handleSubmit={handleSubmit}
           />
         </div>
@@ -96,4 +92,4 @@ function TaskPreviewView({ description }) {
   );
 }
 
-ReactDOM.render(<RemoteQueryApp />, document.getElementById("app"));
+ReactDOM.render(<RemoteProcedureApp />, document.getElementById("app"));
