+++ examples/remote_procedure/mnist/webapp/src/components/core_components.jsx
