+++ examples/remote_procedure/mnist/webapp/src/css/style.css
