+++ examples/remote_procedure/mnist/webapp/src/main.js
