+++ examples/remote_procedure/mnist/webapp/src/static/index.html
