+++ examples/remote_procedure/mnist/webapp/webpack.config.js
