+++ examples/remote_procedure/template/README.md
@@ -1,6 +1,6 @@
 # Mephisto Remote Query Template
 
-This task is an _extremely rough_ bare-bones example of being able to set up a task where the frontend directly connects to the backend using the `mephisto-remote-query` package. It should serve as a decent example of how to get a `RemoteQueryBlueprint` task up off the ground.
+This task is an _extremely rough_ bare-bones example of being able to set up a task where the frontend directly connects to the backend using the `mephisto-remote-query` package. It should serve as a decent example of how to get a `RemoteProcedureBlueprint` task up off the ground.
 
 Deploying the task as-is brings up a page where the user needs to click the "query backend" button enough times for the task to be considered ready to submit.
