+++ examples/remote_procedure/template/__init__.py
