+++ examples/remote_procedure/template/webapp/.babelrc
