+++ examples/remote_procedure/template/webapp/src/components/core_components.jsx
