+++ examples/remote_procedure/template/webapp/src/main.js
