+++ examples/remote_procedure/template/webapp/src/static/index.html
