+++ examples/remote_procedure/template/webapp/webpack.config.js
